Merge from devel/network/p2p/deluge:
rev. 13447, by igungor on 2010-08-05
Version bump.

rev. 13472, by igungor on 2010-08-13
Mark the packages whose maintainer declared that he is not maintaining anymore as unmaintained, also fix licenses that are not invalidated and blocks the commit

Package is fixed in devel repository but forgotten to merged to stable. Sorry. Should be fixed.

BUG:FIXED:14038